#0: Split memcpy_to_device related functionality into new memcpy.hpp file
